Hecho el index de productos y parte del editar productos

+++ PHP/proyecto_Servidor/tienda/categorias/index.php
@@ -4,18 +4,23 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Index de tienda</title>
+    <style>
+        .error {
+            color: red;
+        }
+    </style>
     <?php
         error_reporting( E_ALL );
         ini_set( "display_errors", 1 );
         require('../util/conexion.php');
 
-        session_start();
+        /*session_start();
         if (isset($_SESSION["usuario"])) {
             echo "<h2>Bienvenid@ " . $_SESSION["usuario"] .  "</h2>";
         }else{
             header("location: usuario/iniciar_sesion.php");
             exit;
-        }
+        }*/
     ?>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
 </head>
@@ -29,9 +34,23 @@
     <?php
         if($_SERVER["REQUEST_METHOD"] == "POST"){
             $categoria = $_POST["categoria"];
-            //Borrar la categoria
-            $sql = "DELETE FROM categorias WHERE categoria = '$categoria'";
-            $_conexion -> query($sql);
+
+            $sql = "SELECT categoria FROM productos ORDER BY categoria";
+            $resultado = $_conexion -> query($sql);
+            $categorias = [];
+
+            while ($fila = $resultado -> fetch_assoc()) {
+                array_push($categorias, $fila["categoria"]);
+            }
+
+            if (!in_array($categoria, $categorias)) {
+                //Borrar la categoria
+                $sql = "DELETE FROM categorias WHERE categoria = '$categoria'";
+                $_conexion -> query($sql);
+            }
+            else{
+                $err_borrarCategoria = "No se puede borrar la categoria porque esta asociada a un producto";
+            }   
         }
 
         $sql = "SELECT * FROM categorias";
@@ -39,6 +58,7 @@
     ?>
     <a class="btn btn-secondary" href="nueva_categoria.php">Crear nueva categoria</a>
     <br><br>
+    <?php if(isset($err_borrarCategoria)) echo "<span class='error'>$err_borrarCategoria</span>" ?>
     <table class="table table-striped table-hover">
         <thead class="table-dark">
             <tr>
